pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
extend-exclude = ".*\\.ipynb"
target-version = ["py311"]

[tool.mypy]
exclude = "/migrations/"

check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_reexport = true
show_error_codes = true
strict_concatenate = true
strict_equality = true
warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
warn_unused_configs = true

plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "zelda.settings"

[tool.ruff]
select = [
    "A",
    "ARG",
    "B",
    "BLE",
    "C4",
    "DTZ",
    "E",
    "ERA",
    "EXE",
    "F",
    "FBT",
    "G",
    "I",
    "INP",
    "ISC",
    "N",
    "PGH",
    "PIE",
    "PLC",
    "PLE",
    "PLW",
    "PT",
    "PTH",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLF",
    "T10",
    "TID",
    "TRY",
    "UP",
    "W",
]
ignore = [
    "E501",
    "RUF012",
    "TRY003",
]
src = ["src"]
target-version = "py311"

[tool.ruff.per-file-ignores]
"tests/**" = [
    "FBT001",
    "PT011",
    "S101",
    "S106",
]
"**/migrations/**" = [
    "RUF001",
]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.flake8-tidy-imports.banned-api]
"mock".msg = "Use unittest.mock"
"pytz".msg = "Use zoneinfo"

[tool.ruff.isort]
combine-as-imports = true
forced-separate = ["tests"]
split-on-trailing-comma = false

[tool.pytest.ini_options]
addopts = "-vv --reuse-db"
testpaths = "tests"

[tool.coverage.run]
source = ["src/"]
data_file = ".cov_cache/coverage.dat"
omit = [
    "src/zelda/asgi.py",
    "src/zelda/wsgi.py",
    "src/zelda/settings.py",
    "src/zelda/*/migrations/*",
    "src/zelda/lib/workers.py",
    "src/zelda/home/dashboard.py",
    "src/zelda/home/management/commands/makemigrations.py",
    "src/zelda/home/management/commands/checkmigrations.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
skip_empty = true

[tool.poetry]
name = "zelda"
version = "0.1.0"
description = "Zelda armor app"
license = "Proprietary"
authors = [
    "Stephanos Kuma <stephanos@kuma.ai>",
]

[tool.poetry.dependencies]
# python version
python = "^3.11"

# dependencies
dj-settings = "~4.2"
Django = "~4.2"
django-stubs-ext = "~4.2"
django-grappelli = "~3.0"
gunicorn = "~20.1"
pathurl = "~0.6"
psycopg = "~3.1"
pyOpenSSL = "~23.2"
uvicorn = "~0.22"

[tool.poetry.group.dev.dependencies]
django-extensions = "~3.2"
ipdb = "~0.13"
ipython = "~8.14"
pipdeptree = "~2.9"
Werkzeug = "~2.3"

[tool.poetry.group.lint.dependencies]
black = "~23.3"
django-stubs = "~4.2"
mypy = "~1.4"
ruff = "~0.0"

[tool.poetry.group.test.dependencies]
factorio = "~0.2"
freezegun = "~1.2"
pytest = "~7.4"
pytest-asyncio = "~0.21"
pytest-cov = "~4.1"
pytest-django = "~4.5"
